requirements.txt
importlib-metadata==6.8.0
importlib-resources==6.0.0
libusb1==3.0.0
libusb==1.0.26b5
ndeflib==0.3.3
nfcpy==1.0.4
packaging==23.1
pkg-about==1.0.8
pyDes==2.0.1
pyscard==2.0.7
pyserial==3.5
pyusb==1.2.1
tomli==2.0.1
zipp==3.16.2
